Remove special-case bicycles
