+++ src/main/resources/beam-template.conf
@@ -82,7 +82,6 @@ beam.agentsim.agents.plans.inputPersonAttributesFile = ${beam.inputDirectory}"/p
 beam.agentsim.agents.households.inputFile= ${beam.inputDirectory}"/households.xml.gz"
 beam.agentsim.agents.households.inputHouseholdAttributesFile = ${beam.inputDirectory}"/householdAttributes.xml.gz"
 
-beam.agentsim.agents.vehicles.bicycles.useBikes="boolean | false"
 #BeamVehicles Params
 beam.agentsim.agents.vehicles.beamFuelTypesFile = ${beam.inputDirectory}"/beamFuelTypes.csv"
 beam.agentsim.agents.vehicles.beamVehicleTypesFile = ${beam.inputDirectory}"/vehicleTypes.csv"
