+++ src/main/scala/beam/agentsim/agents/vehicles/BeamVehicleType.scala
@@ -52,17 +52,6 @@ object BeamVehicleType {
       vehicleCategory = Body
     )
 
-  val defaultBicycleBeamVehicleType: BeamVehicleType = BeamVehicleType(
-    Id.create("BIKE_TYPE_DEFAULT", classOf[BeamVehicleType]),
-    1,
-    0,
-    1.5,
-    Food,
-    defaultHumanBodyBeamVehicleType.primaryFuelConsumptionInJoulePerMeter / 5.0, // 5x more efficient than walking
-    defaultHumanBodyBeamVehicleType.primaryFuelCapacityInJoule, // same capacity as human body
-    vehicleCategory = Bike
-  )
-
   val powerTrainForHumanBody: Powertrain = new Powertrain(
     BeamVehicleType.defaultHumanBodyBeamVehicleType.primaryFuelConsumptionInJoulePerMeter
   )
