+++ src/main/scala/beam/sim/config/BeamConfig.scala
@@ -1,4 +1,4 @@
-// generated by tscfg 0.9.4 on Wed Jan 23 23:55:31 IST 2019
+// generated by tscfg 0.9.4 on Thu Feb 21 12:37:02 CET 2019
 // source: src/main/resources/beam-template.conf
 
 package beam.sim.config
@@ -535,25 +535,11 @@ object BeamConfig {
           beamFuelTypesFile: java.lang.String,
           beamVehicleTypesFile: java.lang.String,
           beamVehiclesFile: java.lang.String,
-          bicycles: BeamConfig.Beam.Agentsim.Agents.Vehicles.Bicycles,
           sharedFleets: scala.List[BeamConfig.Beam.Agentsim.Agents.Vehicles.SharedFleets$Elm],
           transitVehicleTypesByRouteFile: java.lang.String
         )
 
         object Vehicles {
-          case class Bicycles(
-            useBikes: scala.Boolean
-          )
-
-          object Bicycles {
-
-            def apply(c: com.typesafe.config.Config): BeamConfig.Beam.Agentsim.Agents.Vehicles.Bicycles = {
-              BeamConfig.Beam.Agentsim.Agents.Vehicles.Bicycles(
-                useBikes = c.hasPathOrNull("useBikes") && c.getBoolean("useBikes")
-              )
-            }
-          }
-
           case class SharedFleets$Elm(
             fixed_non_reserving: scala.Option[
               BeamConfig.Beam.Agentsim.Agents.Vehicles.SharedFleets$Elm.FixedNonReserving
@@ -629,10 +615,6 @@ object BeamConfig {
               beamVehiclesFile =
                 if (c.hasPathOrNull("beamVehiclesFile")) c.getString("beamVehiclesFile")
                 else "/test/input/beamville/vehicles.csv",
-              bicycles = BeamConfig.Beam.Agentsim.Agents.Vehicles.Bicycles(
-                if (c.hasPathOrNull("bicycles")) c.getConfig("bicycles")
-                else com.typesafe.config.ConfigFactory.parseString("bicycles{}")
-              ),
               sharedFleets = $_LBeamConfig_Beam_Agentsim_Agents_Vehicles_SharedFleets$Elm(c.getList("sharedFleets")),
               transitVehicleTypesByRouteFile =
                 if (c.hasPathOrNull("transitVehicleTypesByRouteFile")) c.getString("transitVehicleTypesByRouteFile")
