+++ test/input/equil-square/equil-0.001k.conf
@@ -133,8 +133,6 @@ beam.agentsim.agents.rideHail.allocationManager.repositionLowWaitingTimes.produc
 
 beam.agentsim.agents.rideHail.iterationStats.timeBinSizeInSec=3600
 
-# Use bikes?
-beam.agentsim.agents.vehicles.bicycles.useBikes=false
 #BeamVehicles Params
 beam.agentsim.agents.vehicles.beamFuelTypesFile = ${beam.inputDirectory}"/../beamville/beamFuelTypes.csv"
 beam.agentsim.agents.vehicles.beamVehicleTypesFile = ${beam.inputDirectory}"/../beamville/vehicleTypes.csv"
