+++ test/input/sf-light/sf-light-25k.conf
@@ -45,7 +45,6 @@ beam.agentsim.agents.households {
 beam.agentsim.agents.vehicles.beamFuelTypesFile = ${beam.inputDirectory}"/beamFuelTypes.csv"
 beam.agentsim.agents.vehicles.beamVehicleTypesFile = ${beam.inputDirectory}"/vehicleTypes.csv"
 beam.agentsim.agents.vehicles.beamVehiclesFile = ${beam.inputDirectory}"/sample/25k/vehicles.csv.gz"
-beam.agentsim.agents.vehicles.bicycles.useBikes=false
 beam.agentsim.agents.vehicles.sharedFleets = []
 #TAZ params
 beam.agentsim.taz.file=${beam.inputDirectory}"/taz-centers.csv.gz"
